Code quality remediation: 22 findings across 9 files
CRITICAL fixes:
- direct_lsh.c: free pair_ig array and per-pair buffers (memory leak)
- m4t_route.c: replace assert-in-release with fprintf on malloc failure
- tests/CMakeLists.txt: wrap smoke tests in if(GLYPH_BUILD_TOOLS),
  add add_dependencies() for tool targets

HIGH fixes:
- glyph_sig.c: cast both operands to size_t before multiplication in
  4 malloc calls (integer overflow on large dimensions)
- glyph_dataset.c: overflow check on rows*cols dimension product
- m4t_trit_pack.c: add assert(a && b && mask) to m4t_popcount_dist
- CMakeLists.txt: link all 15 research tools against glyph (not m4t
  directly) — glyph transitively provides m4t
- CMakeLists.txt: update stale comment about "one consumer"

MEDIUM fixes:
- direct_lsh.c: extract topk_insert() and topk_vote() helpers,
  replacing 3 copy-pasted insertion-sort blocks (~45 lines removed)
- direct_lsh.c: expand dense entropy calculation for readability
- m4t_route.h: clarify scratch buffer ownership semantics
- glyph_dataset.c: path length validation before snprintf (both
  MNIST and CIFAR-10 loaders)
- m4t_mtfp.c: assert(n >= 0) in m4t_mtfp_vec_zero

+++ src/glyph_sig.c
@@ -60,7 +60,7 @@ int glyph_sig_builder_init(
     glyph_rng_t rng;
     glyph_rng_seed(&rng, s0, s1, s2, s3);
 
-    proj_w = malloc((size_t)n_proj * input_dim);
+    proj_w = malloc((size_t)n_proj * (size_t)input_dim);
     if (!proj_w) goto cleanup;
     for (int i = 0; i < n_proj * input_dim; i++) {
         uint32_t r = glyph_rng_next(&rng) % 3;
@@ -68,14 +68,14 @@ int glyph_sig_builder_init(
     }
 
     int proj_Dp = M4T_TRIT_PACKED_BYTES(input_dim);
-    sb->proj_packed = malloc((size_t)n_proj * proj_Dp);
+    sb->proj_packed = malloc((size_t)n_proj * (size_t)proj_Dp);
     if (!sb->proj_packed) goto cleanup;
     m4t_pack_trits_rowmajor(sb->proj_packed, proj_w, n_proj, input_dim);
 
     /* Calibrate tau from the |projection| distribution of the calibration
      * subset. Uses the same percentile-at-density rule every cascade tool
      * used before this refactor. */
-    calib_proj = malloc((size_t)n_calib * n_proj * sizeof(m4t_mtfp_t));
+    calib_proj = malloc((size_t)n_calib * (size_t)n_proj * sizeof(m4t_mtfp_t));
     if (!calib_proj) goto cleanup;
     for (int i = 0; i < n_calib; i++) {
         m4t_mtfp_ternary_matmul_bt(
